用户处理中添加限制条件,防止胡乱修改导致其他人无法正常登陆,自己运行时可以去掉这部分判断

+++ ssm-maven/src/main/java/com/ssm/maven/core/service/impl/UserServiceImpl.java
@@ -34,6 +34,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public int updateUser(User user) {
+        //防止有人胡乱修改导致其他人无法正常登陆
+        if ("admin".equals(user.getUserName())) {
+            return 0;
+        }
         return userDao.updateUser(user);
     }
 
@@ -49,6 +53,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public int deleteUser(Integer id) {
+        //防止有人胡乱修改导致其他人无法正常登陆
+        if (2 == id) {
+            return 0;
+        }
         return userDao.deleteUser(id);
     }
